+++ src/browseraction/components/ActionBar.vue
@@ -2,7 +2,9 @@
   <section id="action-bar">
     <button id="extract" tabindex="6" @click="setUrlListInputData">Extract URLs from text</button>
     <button id="open" tabindex="2" @click="openURLs">
-      <strong>Open URLs</strong>
+      <strong>
+        Open URLs <span v-if="tabCount > 0">({{ tabCount }})</span>
+      </strong>
     </button>
     <select
       id="tabGroupSelection"
@@ -14,18 +16,17 @@
         {{ group.title }}
       </option>
     </select>
+    <span
+      id="tabcount"
+      v-if="tabCount >= 25"
+      aria-label="Opening many URLs at once may lead to long wait times or crash your browser."
+      data-microtip-position="bottom"
+      data-microtip-size="medium"
+      role="tooltip"
+    >
+      &#9888;
+    </span>
   </section>
-  <span id="tabcount" v-if="tabCount !== '0'">
-    <abbr title="Opening too many tabs at once may lead to long wait times or crash your browser.">
-      &#9432;
-      <span>
-        will open
-        <span id="tabcount-number">{{ tabCount }}</span>
-        new
-        <span id="tabcount-tab-label">tab<span v-if="tabCount !== '1'">s</span></span>
-      </span>
-    </abbr>
-  </span>
 </template>
 
 <script lang="ts">